info

+++ db.php
@@ -10,8 +10,26 @@ function conn(){
    return $con;
 }
 
+
+function check_input(&$value)
+{
+// 去除斜杠
+if (get_magic_quotes_gpc())
+  {
+  $value = stripslashes($value);
+  }
+// 如果不是数字则加引号
+if (!is_numeric($value))
+  {
+  $value = mysql_real_escape_string($value);
+  }
+}
+
+
 function login($username,$password){
     $con=conn();
+    check_input($username);
+    check_input($password);
     $res=mysql_query("select * from user where user_name='".$username."' and user_password='".$password."' ");
     mysql_close($con);
     return $res;
@@ -19,6 +37,7 @@ function login($username,$password){
 
 function getBlogById($id){ 
   $con=conn();
+  check_input($id);
   $res=mysql_query("select * from blog where blog_id='".$id."'");
   mysql_close($con);
   return $res;  
@@ -26,6 +45,7 @@ function getBlogById($id){
 
 function getOldImageUrl($id){
   $con=conn();
+  check_input($id);
   $res=mysql_query("select user_image_url from user_info where user_id ='".$id."'");
   $url=mysql_fetch_row($res);
   mysql_close($con);
@@ -34,6 +54,7 @@ function getOldImageUrl($id){
 
 function searchUserByName($username){
   $con=conn();
+  check_input($username);
   $res=mysql_query("select count(*) from user where user_name='".$username."'");
   $nums=mysql_fetch_row($res);
   mysql_close($con);
@@ -56,6 +77,7 @@ function getHotBlogs(){
 
 function getBlogsByUid($id){
   $con=conn();
+  check_input($id);
   $res=mysql_query("select * from blog where  blog_user_id='".$id."' order by  blog.blog_cretime desc"); 
   mysql_close($con);
   return $res;
@@ -63,6 +85,7 @@ function getBlogsByUid($id){
 
 function getUserinfoById($id){
   $con=conn();
+   check_input($id);
   $res=mysql_query("select * from user_info where user_id='".$id."'");
   $user_info=mysql_fetch_array($res);
   mysql_close($con);
@@ -71,6 +94,7 @@ function getUserinfoById($id){
 
 function getCommentsByBlogId($id){
   $con=conn();
+   check_input($id);
   $res=mysql_query("select count(*) from blog_comment where blog_id='".$id."'");
   $nums=mysql_fetch_row($res);
   mysql_close($con);
@@ -79,6 +103,7 @@ function getCommentsByBlogId($id){
 
 function getCommentByBlogId($id){
   $con=conn();
+   check_input($id);
   $res=mysql_query("select * from blog_comment where blog_id='".$id."' ORDER BY  blog_comment.blog_comment_cretime DESC ");
   mysql_close($con);
   return $res;
@@ -86,6 +111,7 @@ function getCommentByBlogId($id){
 
 function getMessageByUidOld($id){
   $con=conn();
+   check_input($id);
   $res=mysql_query("select * from  user_message where user_message_uid='".$id."' and user_message_state=1 order by  user_message.user_message_cretime desc ");
   mysql_close($con);
   return $res;
@@ -93,6 +119,7 @@ function getMessageByUidOld($id){
 
 function getMessageByUidNew($id){
   $con=conn();
+   check_input($id);
   $res=mysql_query("select * from  user_message where user_message_uid='".$id."' and user_message_state=0 order by  user_message.user_message_cretime desc ");
   mysql_close($con);
   return $res;
@@ -100,6 +127,7 @@ function getMessageByUidNew($id){
 
 function getMessagesCount($id){
   $con=conn();
+   check_input($id);
   $res=mysql_query("select count(*) from  user_message where user_message_uid='".$id."' and user_message_state=0");
   $num=mysql_fetch_row($res);
   mysql_close($con);
@@ -108,6 +136,7 @@ function getMessagesCount($id){
 
 function getImageByUid($id){
   $con=conn();
+   check_input($id);
   $res=mysql_query("select user_image_url from user_info where user_id ='".$id."' ");
   $num=mysql_fetch_row($res);
   mysql_close($con);
@@ -116,6 +145,8 @@ function getImageByUid($id){
 
 function addLikesByBlogId($id,$m_uid){
   $con=conn();
+  check_input($id);
+  check_input($m_uid);
   $res=mysql_query("select * from blog where blog_id ='".$id."'");
   $blog=mysql_fetch_array($res);
   $res2=mysql_query("select blog_likes_id from blog_likes where blog_id='".$id."' and user_id='".$m_uid."' ");
@@ -139,6 +170,7 @@ function addLikesByBlogId($id,$m_uid){
 
 function  addLikesByCommentId($id){
   $con=conn();
+  check_input($id);
   $res=mysql_query("select blog_comment_likes from blog_comment where blog_comment_id ='".$id."'");
   $nums=mysql_fetch_row($res);
   $nums[0]=$nums[0]+1;
@@ -149,6 +181,7 @@ function  addLikesByCommentId($id){
 
 function getCommentComs($id){
   $con=conn();
+  check_input($id);
   $res=mysql_query("select * from  blog_comment where blog_comment_reply='".$id."'");
   mysql_close($con);
   return $res;
@@ -156,6 +189,7 @@ function getCommentComs($id){
 
 function getCommentComsCount($id){
   $con=conn();
+  check_input($id);
   $res=mysql_query("select count(*) from  blog_comment where blog_comment_reply='".$id."'");
   $nums=mysql_fetch_row($res);
   mysql_close($con);
@@ -164,6 +198,10 @@ function getCommentComsCount($id){
 
 function addUser($username,$password,$email,$sex){
   $con=conn();
+  check_input($username);
+  check_input($password);
+  check_input($email);
+  check_input($sex);
   mysql_query("insert into user(user_name,user_password) values ('".$username."','".$password."')");
   $res=mysql_query("select user_id from user where user_name='".$username."'");
   $user=mysql_fetch_array($res);
@@ -175,6 +213,9 @@ function addUser($username,$password,$email,$sex){
 
 function  addBlog($uid,$title,$content){
   $con=conn();
+  check_input($id);
+  check_input($title);
+  check_input($content);
   $res=mysql_query("insert into blog(blog_user_id,blog_title,blog_content) values('".$uid."','".$title."','".$content."')");
   mysql_close($con);
   if($res > 0){
@@ -186,6 +227,9 @@ function  addBlog($uid,$title,$content){
 
 function alertBlog($blog_id,$title,$content){
   $con=conn();
+  check_input($blog_id);
+  check_input($title);
+  check_input($content);
   $res=mysql_query("update blog set blog_title='".$title."',blog_content='".$content."' where blog_id='".$blog_id."' ");
   mysql_close($con);
   if($res > 0){
@@ -198,6 +242,9 @@ function alertBlog($blog_id,$title,$content){
 
 function addBlogComment($uid,$comment,$blog_id){
   $con=conn();
+  check_input($id);
+  check_input($comment);
+  check_input($blog_id);
   $res=mysql_query("insert into blog_comment(bolg_comment_uid,blog_comment_content,blog_id) values('".$uid."','".$comment."','".$blog_id."')");
   $res2=mysql_query("select blog_user_id from blog where blog_id='".$blog_id."'");
   $id=mysql_fetch_row($res2);
@@ -214,6 +261,10 @@ function addBlogComment($uid,$comment,$blog_id){
 
 function addCommentComs($uid,$comment_id,$commentComs,$reUid){
   $con=conn();
+  check_input($id);
+  check_input($comment_id);
+  check_input($commentComs);
+  check_input($reUid);
   $res=mysql_query("insert into blog_comment(bolg_comment_uid,blog_comment_content,blog_comment_reply,blog_comment_reUid) values('".$uid."','".$commentComs."','".$comment_id."','".$reUid."')");
   $res2=mysql_query("select blog_id from blog_comment where blog_comment_id='".$comment_id."'");
   $blog_id=mysql_fetch_row($res2);
@@ -230,6 +281,12 @@ function addCommentComs($uid,$comment_id,$commentComs,$reUid){
 
 function updateUserInfo($uid,$nickname,$email,$age,$sex,$introduce){
   $con=conn();
+  check_input($uid);
+  check_input($nickname);
+  check_input($email);
+  check_input($age);
+  check_input($sex);
+  check_input($introduce);
   $res=mysql_query("update user_info set user_info_nickname='".$nickname."',user_sex='".$sex."',user_age='".$age."',user_mail='".$email."',user_introduce='".$introduce."' where user_id='".$uid."'");
   mysql_close($con);
   if($res>0){
@@ -241,6 +298,7 @@ function updateUserInfo($uid,$nickname,$email,$age,$sex,$introduce){
 
 function updeteMessageById($id){
   $con=conn();
+  check_input($id);
   mysql_query("update user_message set user_message_state=1 where user_message_uid='".$id."'");
   mysql_close($con);
   return true;
@@ -248,16 +306,37 @@ function updeteMessageById($id){
 
 function updateImageById($id,$dst){
   $con=conn();
+  check_input($id);
+  check_input($dst);
   mysql_query("update user_info set user_image_url='".$dst."' where user_id='".$id."' ");
   mysql_close($con);
   return true;
 }
 
-function delBlogById($id){
+function delBlogById($id,$uid){
   $con=conn();
-  mysql_query("delete from blog where  blog_id ='".$id."' ");
+  check_input($id);
+  check_input($uid);
+  $res= mysql_query("delete from blog where  blog_id ='".$id."' and blog_user_id='".$uid."' ");
   mysql_close($con);
-  return true; 
+  if($res >0){
+      return true; 
+  }else{
+      return false;
+  }
+}
+
+function delCommentById($id,$uid){
+   $con=conn();
+   check_input($id);
+   check_input($uid);
+   $res=mysql_query("delete from blog_comment where  blog_comment_id ='".$id."' and bolg_comment_uid='".$uid."' ");
+   mysql_close($con);
+   if($res >0){
+      return true; 
+  }else{
+      return false;
+  }
 }
 
 ?>
